+++ src/web_gravitymon.cpp
@@ -52,12 +52,13 @@ constexpr auto PARAM_FEATURE_CHARGING_SUPPORTED = "charging";
 
 void GravitymonWebServer::doWebCalibrateStatus(JsonObject &obj) {
   if (myGyro.isConnected()) {
-    if(_gyroCalibrationSuccess) {
+    if (_gyroCalibrationSuccess) {
       obj[PARAM_SUCCESS] = true;
       obj[PARAM_MESSAGE] = "Calibration completed";
     } else {
       obj[PARAM_SUCCESS] = false;
-      obj[PARAM_MESSAGE] = "Calibration failed, faulty gyro or movement during calibration";
+      obj[PARAM_MESSAGE] =
+          "Calibration failed, faulty gyro or movement during calibration";
     }
   } else {
     obj[PARAM_SUCCESS] = false;
@@ -150,7 +151,8 @@ void GravitymonWebServer::doTaskSensorCalibration() {
 
   if (myGyro.isConnected()) {
     _gyroCalibrationSuccess = myGyro.calibrateSensor();
-    Log.info(F("WEB : Gyro calibration %s" CR), _gyroCalibrationSuccess ? "succeeded" : "failed");
+    Log.info(F("WEB : Gyro calibration %s" CR),
+             _gyroCalibrationSuccess ? "succeeded" : "failed");
   } else {
     Log.error(F("WEB : No gyro connected, skipping calibration" CR));
   }